Use header-based vku.
