+++ impl.cpp
@@ -1,2 +1,7 @@
+#define VMA_IMPLEMENTATION
+#include <vk_mem_alloc.h>
+
+#define STB_IMAGE_IMPLEMENTATION
+#include <stb_image.h>
 #define STB_IMAGE_WRITE_IMPLEMENTATION
 #include <stb_image_write.h>
